Rename di to dc SDK
