+++ sdk/dc/build.gradle.kts
