+++ sdk/dc/proguard-rules.pro
