+++ sdk/dc/src/main/AndroidManifest.xml
