--- sdk/di/src/main/java/com/ibm/security/verifysdk/di/QrCode.kt
@@ -1,13 +1,0 @@
-/*
- * Copyright contributors to the IBM Security Verify SDK for Android project
- */
-
-package com.ibm.security.verifysdk.di
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class QrCode(
-    var type : String? = null,
-    var data : QrCodeData?   = QrCodeData()
-)
