+++ sdk/examples/dc_demo/.gitignore
