+++ sdk/examples/dc_demo/build.gradle.kts
@@ -1,0 +1,54 @@
+plugins {
+    alias(libs.plugins.android.application)
+    alias(libs.plugins.jetbrains.kotlin.android)
+    alias(libs.plugins.kotlin.serialization)
+}
+
+apply(from = "$rootDir/common-config-demos.gradle")
+apply(from = "$rootDir/common-config-ktor.gradle")
+
+android {
+    namespace = "com.ibm.security.verifysdk.dc.demoapp"
+    defaultConfig {
+        applicationId = "com.ibm.security.verifysdk.dc.demoapp"
+
+        manifestPlaceholders["auth_redirect_scheme"] = "verifysdk"
+        manifestPlaceholders["auth_redirect_host"] = "callback"
+        manifestPlaceholders["auth_redirect_path"] = "/redirect"
+    }
+
+    buildFeatures {
+        compose = true
+    }
+    composeOptions {
+        kotlinCompilerExtensionVersion = "1.5.1"
+    }
+    packaging {
+        resources {
+            excludes += "/META-INF/{AL2.0,LGPL2.1}"
+        }
+    }
+
+}
+
+dependencies {
+
+    implementation(project(":di"))
+    implementation(project(":core"))
+    implementation(project(":authentication"))
+
+    implementation(libs.androidx.material3)
+    implementation(libs.androidx.ui)
+    implementation(libs.androidx.ui.graphics)
+    implementation(libs.androidx.ui.tooling.preview)
+    implementation(libs.zxing.android.embedded)
+    implementation(platform(libs.androidx.compose.bom))
+
+    androidTestImplementation(libs.androidx.espresso.core)
+    androidTestImplementation(libs.androidx.junit)
+    androidTestImplementation(libs.androidx.ui.test.junit4)
+    androidTestImplementation(platform(libs.androidx.compose.bom))
+    debugImplementation(libs.androidx.ui.test.manifest)
+    debugImplementation(libs.androidx.ui.tooling)
+    testImplementation(libs.junit)
+}
