+++ sdk/examples/dc_demo/proguard-rules.pro
