+++ sdk/examples/dc_demo/src/main/AndroidManifest.xml
