+++ sdk/examples/dc_demo/src/main/res/drawable/ic_launcher_foreground.xml
