+++ sdk/examples/dc_demo/src/main/res/mipmap-anydpi/ic_launcher_round.xml
