+++ sdk/examples/dc_demo/src/main/res/raw/trusted_roots.pem
