+++ sdk/examples/dc_demo/src/main/res/values/themes.xml
