+++ sdk/examples/dc_demo/src/main/res/xml/network_security_config.xml
