+++ sdk/settings.gradle.kts
@@ -23,16 +23,16 @@ rootProject.name = "IBM Security Verify SDK"
 include(":adaptive")
 include(":authentication")
 include(":core")
-include(":di")
+include(":dc")
 include(":dpop_demo")
 include(":fido2")
 include(":fido2_demo")
 include(":mfa")
 include(":mfa_demo")
-include(":di_demo")
+include(":dc_demo")
 
 project(":fido2_demo").projectDir = File("examples/fido2_demo")
 project(":dpop_demo").projectDir = File("examples/dpop_demo")
 project(":mfa_demo").projectDir = File("examples/mfa_demo")
-project(":di_demo").projectDir = File("examples/di_demo")
+project(":dc_demo").projectDir = File("examples/dc_demo")
